% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_input.R
\name{convert_image}
\alias{convert_image}
\title{Process an image file to text.}
\usage{
convert_image(image_file, image_format)
}
\arguments{
\item{image_file}{A image file of JPG, BMP, or TIFF format imported using readJPG or similar.}

\item{image_format}{Image format as a string.}
}
\value{
A processed numeric matrix containing image information.
}
\description{
Process an image file to text.
}
\examples{
data(G080_image)
convert_image(G080_image, "bmp")
}
