% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{findcontrast}
\alias{findcontrast}
\title{Find an overall contrast of intensity values in the general neighborhood of a minutiae.}
\usage{
findcontrast(kk, jj, print_info)
}
\arguments{
\item{kk}{X pixel value of a given feature.}

\item{print_info}{Fingerprint image as matrix.}

\item{yy}{Y pixel value of a given feature.}
}
\value{
The largest intensity difference between a point and any neighbor intensity in a 3x3 region, divided by the intensity in the print (usually 255).
}
\description{
Find an overall contrast of intensity values in the general neighborhood of a minutiae.
}
\examples{
print_image <- convert_image(G080_image, "bmp")
findcontrast(831, 542, print_image)
}
