% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{findmaxgradinregion}
\alias{findmaxgradinregion}
\title{Find the gradient of intensity between the dark and light regions defining a minutiae point.}
\usage{
findmaxgradinregion(x1, y1, areasize, print_info)
}
\arguments{
\item{x1}{X pixel value of a given feature.}

\item{y1}{Y pixel value of a given feature.}

\item{print_info}{Fingerprint image as matrix.}
}
\value{
The maximum gradient within a radius around a given point and the shifted X and Y pixel location values of the pixel in a small neighborhood of the minutiae producing the highest intensity gradient values.
}
\description{
Find the gradient of intensity between the dark and light regions defining a minutiae point.
}
\examples{
print_image <- convert_image(G080_image, "bmp")
findmaxgradinregion(841, 541, print_image)
}
