% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{G080_min}
\alias{G080_min}
\title{G080 minutiae information.}
\format{A matrix with 21 minutiae pixel locations.
\describe{
\item{X}{X pixel value location}
\item{Y}{Y pixel value location}
}}
\usage{
G080_min
}
\description{
X and Y pixel locations for latent G080.
}
\examples{
  G080_min
}
\keyword{datasets}
