% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{quality_scores}
\alias{quality_scores}
\title{Find the quality scores of individual minutiae.}
\usage{
quality_scores(print_info, minutiae_info, verbose = FALSE)
}
\arguments{
\item{print_info}{Fingerprint image as matrix.}

\item{minutiae_info}{Feature (minutiae) information as matrix}

\item{verbose}{If TRUE, will output gradient and contrast information in addition to the quality score.}
}
\value{
Matrix containing the columns: x pixel location of minutiae, y pixel location of minutiae, shifted x location, shifted y location, quality score.
}
\description{
Find the quality scores of individual minutiae.
}
\examples{
data(G080_image); data(G080_min)
print_image <- convert_image(G080_image, "bmp")
min_info <- import_features(G008_min)
quality_scores(print_image, min_info)
}
