% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_input.R
\name{import_features}
\alias{import_features}
\title{Process an file of features (minutiae). Currently assumes LQM input format (columns, in order: minutiae type, x location, y location, ...).}
\usage{
import_features(feature_file)
}
\arguments{
\item{feature_file}{Matrix or data frame containing feature information.}
}
\value{
A processed numeric matrix containing the x and y pixel locations of each minutiae.
}
\description{
Process an file of features (minutiae). Currently assumes LQM input format (columns, in order: minutiae type, x location, y location, ...).
}
\examples{
data(G080_min)
import_features(G080_min)
}
